refactor: split models into separate files and updated references while removing deprecated comments
