+++ tests/test_services.py
@@ -2,8 +2,9 @@ from datetime import timedelta, date
 import pytest
 from fastapi import HTTPException
 
-from app.models.models import TimePointData
-from app.services import time_point_calculation, DateModel, call_weather_api, isolate_day_data, clean_weather_data
+from app.models.weather_models import TimePointData
+from app.services.services import time_point_calculation, call_weather_api, isolate_day_data, clean_weather_data, \
+    DateModel
 
 
 @pytest.mark.parametrize("days_from_today, expected_result", [
@@ -122,23 +123,3 @@ def test_date_model_allows_dates_within_range(good_request_dates):
 def test_data_model_dates_catches_out_of_range_dates(bad_request_dates):
     with pytest.raises(HTTPException):
         DateModel.parse_obj({"date": bad_request_dates})
-
-# @strawberry.type
-# class Query:
-#     @strawberry.field
-#     async def create_user(self, uid: str) -> Dict[str, any]:
-#         matching_list = [user for user in users if uid == user]
-#         if len(matching_list) == 1:
-#             user = users[uid]
-#             return {"message": {uid: user}}
-#         else:
-#             return {"message": f"User, {uid}, does not exist."}
-#
-#
-# schema = strawberry.Schema(Query)
-# graphql_app = GraphQLRouter(schema)
-#
-# app = FastAPI()
-# app.include_router(graphql_app, prefix="/graphql")
-
-#############
